ui: rename tx -> tx details
