+++ src/comps/account/Dashboard.js
@@ -28,7 +28,7 @@ export default props => {
       </div>
 
       <div className="mj-box flex-column w-s">
-        <div><strong>Balance</strong>: mj<strong>{fm(props.userDoc.balance)}</strong> (${fm(props.userDoc.balance * props.mjMetaDoc.val)})</div>
+        <div><strong>Balance</strong>: mj{fm(props.userDoc.balance)}</div>
         <div><strong>Address</strong>: <small>{props.user.uid}</small></div>
         <div><Link to="/receive"><img width="72" src={props.acctQr} alt={props.user.uid}/></Link></div>
       </div>
